pyproject.toml
[tool.poetry]
name = "pyqupa"
version = "0.1.0"
description = "PYthon Qualdich.de mountain PAss data processor."
authors = ["Kyoungseoun Chung <kchung@student.ethz.ch>"]

[tool.poetry.dependencies]
python = "~3.9"
folium = "^0.12.1"
bs4 = "^0.0.1"
lxml = "^4.9.1"
rich = "^12.5.1"
streamlit = "^1.12.2"
numpy = "^1.23.2"
tinydb = "^4.7.0"
streamlit-folium = "^0.6.15"
pydeck = "^0.7.1"
python-dotenv = "^0.20.0"
plotly = "^5.10.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"


[tool.commitizen]
name = "cz_conventional_commits"
version = "0.0.1"
tag_format = "$version"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-x -v -s"
